Fix package name
