+++ libs/finagger/src/main/scala/io/github/cactacea/finagger/Redoc.scala
@@ -1,0 +1,6 @@
+package io.github.cactacea.finagger
+
+import com.twitter.finatra.http.response.Mustache
+
+@Mustache("redoc")
+case class Redoc(modelPath: String)
