+++ libs/finagger/src/main/scala/io/github/cactacea/finagger/WebjarsController.scala
@@ -1,4 +1,4 @@
-package io.cactacea.finagger
+package io.github.cactacea.finagger
 
 import java.util.Date
 import java.util.concurrent.TimeUnit
